sistema-integral-urgencias/requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
prophet==1.1.5
numpy==1.26.2
scikit-learn==1.3.2
